fix docs generation and version parsing; fixes #848
